backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
websockets==13.1
pydantic==2.10.2
python-dotenv==1.0.1
swarms==6.7.7
asyncio==3.4.3
redis==5.2.0
aioredis==2.0.1
httpx==0.28.1
pycoingecko==3.1.0
ccxt==4.4.34
pandas==2.2.3
numpy==2.1.3
ta-lib==0.4.29
python-multipart==0.0.12